Track egg for ultra
This adds a row to track whether the user prefers to use lucky eggs when
becoming ultra friends. I may add a row for best friend later, but
that's kind of out of the realm of PVP.

This also separates the concept of a user from a server, so that in the
future you can register an existing user with a second (third, etc) server
without giving all of the basic information again. This may have
introduced a bug where you cannot register on a second server, now that
I think about it.

It also introduces a breaking api change which technically should be a
major version bump but I'm fudging the rules a bit. (`/list` is now
`player/list`)