Migrations
After some experimenting with the best way to model friendships I went
with this option, which has `user_id` and `friend_id` columns. To ensure
that A doesn't add B and then B adds A, they have to be inserted with
the lower id first. This is not intended for use in the code.

To get a list of users, this is the query I'm planning:

```sql
SELECT fid FROM (
    SELECT user_id as uid, friend_id as fid FROM pvp_user_friendship
    UNION ALL
    SELECT user_id as fid, friend_id as uid FROM pvp_user_friendship
) puf WHERE uid= ?
```

I was going to put it as a view, but as it contains a `UNION` it won't
automatically update. I'm open to redoing this in the future.